requirements.txt
Markdown==3.5
PyYAML>=6.0.1
certifi>=2023.7.22
charset-normalizer>=3.3.0
idna>=3.4
importlib-metadata>=6.8.0
pip>=22.2.2
requests>=2.31.0
setuptools>=65.3.0
tkinterhtml>=0.7
urllib3>=2.0.6
wheel>=0.37.1
zipp>=3.17.0
Pillow~=10.1.0